[Mod] Compute filter cutoff frequency as a float instead of rounded uint32. Impulse Tracker does the same, and the differences are small enough not to warrant a compatibility flag (...hopefully).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@18658 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/SampleFormatSFZ.cpp
@@ -197,9 +197,9 @@ protected:
 
 	static double Identity(double v) noexcept { return v; }
 
-	static double CentsToFilterCutoff(double v, const CSoundFile &sndFile, int envBaseCutoff, uint32 envBaseFreq)
+	static double CentsToFilterCutoff(double v, const CSoundFile &sndFile, int envBaseCutoff, float envBaseFreq)
 	{
-		const auto freq = envBaseFreq * std::pow(2.0, v / 1200.0);
+		const auto freq = static_cast<double>(envBaseFreq) * std::pow(2.0, v / 1200.0);
 		return Util::muldivr(sndFile.FrequencyToCutOff(freq), 127, envBaseCutoff) / 127.0;
 	}
